feat - notFoundException(message) in methods using the ParamConverter

+++ src/Controller/Backoffice/CategoryController.php
@@ -61,8 +61,11 @@ class CategoryController extends AbstractController
     /**
      * @Route("/{id}", name="app_backoffice_category_show", methods={"GET"})
      */
-    public function show(Category $category): Response
+    public function show(?Category $category): Response
     {
+        if (!$category) {
+            throw $this->createNotFoundException("La catégorie demandée n'a pas été trouvée");}
+
         return $this->render('backoffice/category/show.html.twig', [
             'category' => $category,
         ]);
@@ -71,8 +74,11 @@ class CategoryController extends AbstractController
     /**
      * @Route("/{id}/edit", name="app_backoffice_category_edit", methods={"GET", "POST"})
      */
-    public function edit(Request $request, Category $category, CategoryRepository $categoryRepository, CustomSlugger $customSlugger): Response
+    public function edit(Request $request, ?Category $category, CategoryRepository $categoryRepository, CustomSlugger $customSlugger): Response
     {
+        if (!$category) {
+            throw $this->createNotFoundException("La catégorie demandée n'a pas été trouvée");}
+
         $form = $this->createForm(CategoryType::class, $category);
         $form->handleRequest($request);
 
@@ -99,8 +105,11 @@ class CategoryController extends AbstractController
     /**
      * @Route("/{id}", name="app_backoffice_category_delete", methods={"POST"})
      */
-    public function delete(Request $request, Category $category, CategoryRepository $categoryRepository): Response
+    public function delete(Request $request, ?Category $category, CategoryRepository $categoryRepository): Response
     {
+        if (!$category) {
+            throw $this->createNotFoundException("La catégorie demandée n'a pas été trouvée");}
+
         if ($this->isCsrfTokenValid('delete' . $category->getId(), $request->request->get('_token'))) {
             $categoryRepository->remove($category, true);
         }
