+++ src/Controller/Backoffice/OfferController.php
@@ -20,8 +20,11 @@ class OfferController extends AbstractController
     /**
      * @Route("/{id}", name="app_backoffice_offer_show", methods={"GET"})
      */
-    public function show(Offer $offer): Response
+    public function show(?Offer $offer): Response
     {
+        if (!$offer) {
+            throw $this->createNotFoundException("L'offre demandée n'a pas été trouvée");}
+
         return $this->render('backoffice/offer/show.html.twig', [
             'offer' => $offer,
         ]);
@@ -30,8 +33,11 @@ class OfferController extends AbstractController
     /**
      * @Route("/{id}/edit", name="app_backoffice_offer_edit", methods={"GET", "POST"})
      */
-    public function edit(Request $request, Offer $offer, OfferRepository $offerRepository): Response
+    public function edit(Request $request, ?Offer $offer, OfferRepository $offerRepository): Response
     {
+        if (!$offer) {
+            throw $this->createNotFoundException("L'offre demandée n'a pas été trouvée");}
+
         $form = $this->createForm(OfferType::class, $offer);
         $form->handleRequest($request);
 
@@ -55,8 +61,11 @@ class OfferController extends AbstractController
     /**
      * @Route("/{id}", name="app_backoffice_offer_delete", methods={"POST"})
      */
-    public function delete(Request $request, Offer $offer, OfferRepository $offerRepository): Response
+    public function delete(Request $request, ?Offer $offer, OfferRepository $offerRepository): Response
     {
+        if (!$offer) {
+            throw $this->createNotFoundException("L'offre demandée n'a pas été trouvée");}
+
         if ($this->isCsrfTokenValid('delete' . $offer->getId(), $request->request->get('_token'))) {
             $offerRepository->remove($offer, true);
 
@@ -69,16 +78,17 @@ class OfferController extends AbstractController
     /**
      * @Route("/{id}/validate", name="app_backoffice_offer_validate", methods={"POST"})
      */
-    public function validate(Request $request, Offer $offer, OfferRepository $offerRepository): Response
+    public function validate(Request $request, ?Offer $offer, OfferRepository $offerRepository): Response
     {
-        // if ($this->isCsrfTokenValid('validate' . $offer->getId(), $request->request->get('_token'))) {
-        // }
-        
+        if ($this->isCsrfTokenValid('validate' . $offer->getId(), $request->request->get('_token'))) {
+        }
+        if (!$offer) {
+            throw $this->createNotFoundException("L'offre demandée n'a pas été trouvée");}
         $offer->setIsReported(false);
         $offer->setUpdatedAt(new DateTime());
         $offerRepository->add($offer, true);
 
 
-        return $this->redirectToRoute('app_backoffice_reported', [], Response::HTTP_SEE_OTHER);
+        return $this->redirectToRoute('app_backoffice_reported_index', [], Response::HTTP_SEE_OTHER);
     }
 }
