+++ src/Controller/Backoffice/UserController.php
@@ -62,8 +62,11 @@ class UserController extends AbstractController
     /**
      * @Route("/{id}", name="app_backoffice_user_show", methods={"GET"})
      */
-    public function show(User $user): Response
+    public function show(?User $user): Response
     {
+        if (!$user) {
+            throw $this->createNotFoundException("L'utilisateur demandé n'a pas été trouvé");}
+
         return $this->render('backoffice/user/show.html.twig', [
             'user' => $user,
         ]);
@@ -72,9 +75,12 @@ class UserController extends AbstractController
     /**
      * @Route("/{id}/edit", name="app_backoffice_user_edit", methods={"GET", "POST"})
      */
-    public function edit(Request $request, User $user, UserRepository $userRepository): Response
+    public function edit(Request $request, ?User $user, UserRepository $userRepository): Response
     {
         $this->denyAccessUnlessGranted("ROLE_ADMIN");
+
+        if (!$user) {
+            throw $this->createNotFoundException("L'utilisateur demandé n'a pas été trouvé");}
         
         $form = $this->createForm(UserTypeEdit::class, $user);
         if(!in_array('ROLE_ADMIN', $this->getUser()->getRoles())) {
@@ -104,9 +110,12 @@ class UserController extends AbstractController
     /**
      * @Route("/{id}", name="app_backoffice_user_delete", methods={"POST"})
      */
-    public function delete(Request $request, User $user, UserRepository $userRepository): Response
+    public function delete(Request $request, ?User $user, UserRepository $userRepository): Response
     {
         $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
+        if (!$user) {
+            throw $this->createNotFoundException("L'utilisateur demandé n'a pas été trouvé");}
 
         if ($this->isCsrfTokenValid('delete' . $user->getId(), $request->request->get('_token'))) {
             $userRepository->remove($user, true);
