+++ src/Controller/Backoffice/WishController.php
@@ -20,8 +20,11 @@ class WishController extends AbstractController
     /**
      * @Route("/{id}", name="app_backoffice_wish_show", methods={"GET"})
      */
-    public function show(Wish $wish): Response
+    public function show(?Wish $wish): Response
     {
+        if (!$wish) {
+            throw $this->createNotFoundException("La demande demandée n'a pas été trouvée");}
+
         return $this->render('backoffice/wish/show.html.twig', [
             'wish' => $wish,
         ]);
@@ -30,8 +33,11 @@ class WishController extends AbstractController
     /**
      * @Route("/{id}/edit", name="app_backoffice_wish_edit", methods={"GET", "POST"})
      */
-    public function edit(Request $request, Wish $wish, WishRepository $wishRepository): Response
+    public function edit(Request $request, ?Wish $wish, WishRepository $wishRepository): Response
     {
+        if (!$wish) {
+            throw $this->createNotFoundException("La demande demandée n'a pas été trouvée");}
+
         $form = $this->createForm(WishType::class, $wish);
         $form->handleRequest($request);
 
@@ -55,13 +61,16 @@ class WishController extends AbstractController
     /**
      * @Route("/{id}", name="app_backoffice_wish_delete", methods={"POST"})
      */
-    public function delete(Request $request, Wish $wish, WishRepository $wishRepository): Response
+    public function delete(Request $request, ?Wish $wish, WishRepository $wishRepository): Response
     {
+        if (!$wish) {
+            throw $this->createNotFoundException("La demande demandée n'a pas été trouvée");}
+
         if ($this->isCsrfTokenValid('delete' . $wish->getId(), $request->request->get('_token'))) {
             $wishRepository->remove($wish, true);
         }
 
-        return $this->redirectToRoute('app_backoffice_reported', [], Response::HTTP_SEE_OTHER);
+        return $this->redirectToRoute('app_backoffice_reported_index', [], Response::HTTP_SEE_OTHER);
     }
 
     /**
@@ -69,14 +78,17 @@ class WishController extends AbstractController
      */
     public function validate(Request $request, Wish $wish, WishRepository $wishRepository): Response
     {
-        // if ($this->isCsrfTokenValid('validate' . $wish->getId(), $request->request->get('_token'))) {
-        // }
+        if (!$wish) {
+            throw $this->createNotFoundException("La demande demandée n'a pas été trouvée");}
+            
+        if ($this->isCsrfTokenValid('validate' . $wish->getId(), $request->request->get('_token'))) {
+        }
         
         $wish->setIsReported(false);
         $wish->setUpdatedAt(new DateTime());
         $wishRepository->add($wish, true);
 
 
-        return $this->redirectToRoute('app_backoffice_reported', [], Response::HTTP_SEE_OTHER);
+        return $this->redirectToRoute('app_backoffice_reported_index', [], Response::HTTP_SEE_OTHER);
     }
 }
